[new_bindings] Rename ElunaBinding.h to BindingMap.h
The file is all about the BindingMap class, so I thought the name was more
suiting.
