+++ GameObjectHooks.cpp
@@ -7,7 +7,7 @@
 #include "Hooks.h"
 #include "HookHelpers.h"
 #include "LuaEngine.h"
-#include "ElunaBinding.h"
+#include "BindingMap.h"
 #include "ElunaIncludes.h"
 #include "ElunaEventMgr.h"
 #include "ElunaTemplate.h"
