+++ LuaEngine.cpp
@@ -6,7 +6,7 @@
 
 #include "Hooks.h"
 #include "LuaEngine.h"
-#include "ElunaBinding.h"
+#include "BindingMap.h"
 #include "ElunaEventMgr.h"
 #include "ElunaIncludes.h"
 #include "ElunaTemplate.h"
